Located all em mocks in one place.
