+++ applications/ExecutionManager/CMakeLists.txt
@@ -6,16 +6,13 @@ project (em VERSION 1.0
 file(GLOB_RECURSE EM_SRC src/*.cpp)
 list(REMOVE_ITEM EM_SRC ${CMAKE_CURRENT_SOURCE_DIR}/src/em_main.cpp)
 
-set(EM_INCLUDE ${CMAKE_CURRENT_SOURCE_DIR}/include)
-
 add_subdirectory(test)
 
 add_executable(em src/em_main.cpp ${EM_SRC})
 
 target_link_libraries(em
-    PRIVATE
-      nlohmann_json::nlohmann_json
     PUBLIC
+    nlohmann_json::nlohmann_json
       gen::execution_management
       api::application_state_client
       CapnProto::capnp-rpc
