+++ applications/ExecutionManager/test/CMakeLists.txt
@@ -17,6 +17,7 @@ target_link_libraries(${TEST_NAME}_test
   api::machine_state_client
   gen::execution_management
   CapnProto::capnp-rpc
+  common
 )
 
 target_compile_definitions(${TEST_NAME}_test PUBLIC -DUNIT_TEST)
@@ -44,6 +45,7 @@ target_link_libraries(${COVERAGE_BUILD}
   api::machine_state_client
   gen::execution_management
   CapnProto::capnp-rpc
+  common
 )
 
 target_compile_definitions(${COVERAGE_BUILD} PUBLIC -DUNIT_TEST)
