+++ common/CMakeLists.txt
@@ -13,6 +13,7 @@ target_link_libraries(common
     PUBLIC
       api::application_state_client
       api::machine_state_client
+      nlohmann_json::nlohmann_json
       gmock
 )
 
@@ -22,4 +23,5 @@ target_include_directories(common
       $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
       $<INSTALL_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
       ${MSM_INCLUDE}
+      ${EM_INCLUDE}
 )
